requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-GraphQL==2.0.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
Werkzeug==2.0.3
SQLAlchemy<2.0.0
gunicorn==21.2.0 